+++ tortoise/tests/testmodels.py
@@ -16,7 +16,7 @@ def generate_token():
 
 
 class Tournament(Model):
-    id = fields.IntField(pk=True)
+    id = fields.SmallIntField(pk=True)
     name = fields.CharField(max_length=255)
     desc = fields.TextField(null=True)
     created = fields.DatetimeField(auto_now_add=True, index=True)
@@ -37,7 +37,7 @@ class Reporter(Model):
 
 
 class Event(Model):
-    id = fields.IntField(pk=True)
+    id = fields.BigIntField(pk=True)
     name = fields.TextField()
     tournament = fields.ForeignKeyField("models.Tournament", related_name="events")
     reporter = fields.ForeignKeyField("models.Reporter", null=True)